bats: Do not remove any tests if BATS_TESTS is defined
Otherwise we can't check whether a bug was fixed or not without modifying the code

+++ tests/containers/bats/runc.pm
@@ -54,12 +54,14 @@ sub run {
     # Compile helpers used by the tests
     my $cmds = script_output "find contrib/cmd tests/cmd -mindepth 1 -maxdepth 1 -type d -printf '%f ' || true";
     run_command "make $cmds || true";
-    # Skip this test due to https://github.com/opencontainers/runc/issues/4841
-    run_command "rm -f tests/integration/cgroups.bats" if is_ppc64le;
-    # Skip this test due to https://bugzilla.suse.com/show_bug.cgi?id=1247568
-    run_command "rm -f tests/integration/no_pivot.bats" if is_ppc64le;
-    # Skip this test due to https://bugzilla.suse.com/show_bug.cgi?id=1247567
-    run_command "rm -f tests/integration/seccomp.bats" if is_s390x;
+    unless (get_var("BATS_TESTS")) {
+        # Skip this test due to https://github.com/opencontainers/runc/issues/4841
+        run_command "rm -f tests/integration/cgroups.bats" if is_ppc64le;
+        # Skip this test due to https://bugzilla.suse.com/show_bug.cgi?id=1247568
+        run_command "rm -f tests/integration/no_pivot.bats" if is_ppc64le;
+        # Skip this test due to https://bugzilla.suse.com/show_bug.cgi?id=1247567
+        run_command "rm -f tests/integration/seccomp.bats" if is_s390x;
+    }
 
     my $errors = run_tests(rootless => 1, skip_tests => 'BATS_SKIP_USER');
